requirements.txt
boto3
botocore
requests
flask
pandas
matplotlib
pyyaml
python-dotenv
mysql-connector-python
